+++ .editorconfig
@@ -4,6 +4,9 @@ root = true
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
+indent_style = space
+indent_size = 2
+end_of_line = lf
 
 [*.{js,json,vue}]
 indent_style = space
@@ -12,3 +15,6 @@ indent_size = 2
 [*.sh]
 indent_style = space
 indent_size = 2
+
+[*.md]
+trim_trailing_whitespace = false
