+++ .eslintignore
@@ -1,0 +1,6 @@
+.cache
+.git
+node_modules
+public
+build
+dist
