+++ .prettierignore
@@ -1,0 +1,5 @@
+.cache
+package.json
+package-lock.json
+dist
+node_modules
